refactor: mdx helpers
