+++ src/types/mdx.d.ts
@@ -1,0 +1,11 @@
+export type Frontmatter<T = {}> = {
+  title: string;
+  date: string;
+  description: string;
+} & T;
+
+export type MDXData<T = {}> = {
+  metadata: Frontmatter<T>;
+  slug: string;
+  content: React.ReactNode;
+};
